move JoinExamples to util
